refactor(persona): 插件层退役，角色作为唯一能力绑定入口——技能市场恢复
对照 Codex 源码：codex 无 persona 概念，plugin 纯粹是 skills/MCP 分发层；
LambChat 已有角色广场/发布/复制体系，叠加插件层形成两个能力包概念（重复）。

- 退役插件层：/api/plugins 路由、src/infra/plugin/（含商店→插件迁移）、
  10 个 plugin 错误码、e2e 脚本、启动初始化
- 恢复技能市场：SkillsHub 回到「技能+商店」双 tab，/marketplace 路由还原
- persona 保留 mcp_server_names 绑定 + enabled_mcp_servers 运行时白名单
  全链路（角色直接配置 MCP 的能力不变）；移除 plugin_names 绑定
- persona 编辑器保留「技能 + MCP 服务」两个选择器，移除插件选择器
- 五语 i18n 清理 plugins.* / backendErrors.plugin*（保留 persona MCP 键）
- eager 预算回调 528KB；MCP source_plugin 标记随插件层一并移除
