+++ .gitignore
@@ -112,7 +112,6 @@ YubbChat/
 !/scripts/sync_error_locales.py
 !/scripts/backfill_memory_scope.py
 !/scripts/e2e_local_sandbox.py
-!/scripts/e2e_plugin_hub.py
 !/scripts/generate_updater_manifest.py
 !/scripts/benchmark_memory_scope.py
 docs/docs/.vitepress/
