+++ frontend/src/components/mcp/MCPServerCard.tsx
@@ -79,11 +79,6 @@ export function MCPServerCard({
                 {t("mcp.card.system")}
               </span>
             )}
-            {server.source_plugin && (
-              <span className="scb__status-pill scb__status-pill--installed">
-                {t("mcp.card.fromPlugin", { name: server.source_plugin })}
-              </span>
-            )}
             {!server.enabled && (
               <span className="scb__status-pill scb__status-pill--danger">
                 {t("mcp.card.disabled")}
