+++ frontend/src/components/panels/SkillsHubPanel/__tests__/state.test.ts
@@ -3,7 +3,7 @@ import { resolveSkillsHubTab } from "../state.ts";
 test("keeps the requested tab when both permissions are available", () => {
   expect(resolveSkillsHubTab(undefined, true, true)).toBe("skills");
   expect(resolveSkillsHubTab("skills", true, true)).toBe("skills");
-  expect(resolveSkillsHubTab("plugins", true, true)).toBe("plugins");
+  expect(resolveSkillsHubTab("marketplace", true, true)).toBe("marketplace");
 });
 
 test("resolves to local skills when only local skills are available", () => {
@@ -11,18 +11,18 @@ test("resolves to local skills when only local skills are available", () => {
   expect(resolveSkillsHubTab("skills", true, false)).toBe("skills");
 });
 
-test("resolves to plugins when only plugins are available", () => {
-  expect(resolveSkillsHubTab(undefined, false, true)).toBe("plugins");
-  expect(resolveSkillsHubTab("plugins", false, true)).toBe("plugins");
+test("resolves to marketplace when only marketplace is available", () => {
+  expect(resolveSkillsHubTab(undefined, false, true)).toBe("marketplace");
+  expect(resolveSkillsHubTab("marketplace", false, true)).toBe("marketplace");
 });
 
 test("falls back to the accessible tab when the requested tab is inaccessible", () => {
-  expect(resolveSkillsHubTab("plugins", true, false)).toBe("skills");
-  expect(resolveSkillsHubTab("skills", false, true)).toBe("plugins");
+  expect(resolveSkillsHubTab("marketplace", true, false)).toBe("skills");
+  expect(resolveSkillsHubTab("skills", false, true)).toBe("marketplace");
 });
 
 test("returns null when neither tab is accessible", () => {
   expect(resolveSkillsHubTab(undefined, false, false)).toBe(null);
   expect(resolveSkillsHubTab("skills", false, false)).toBe(null);
-  expect(resolveSkillsHubTab("plugins", false, false)).toBe(null);
+  expect(resolveSkillsHubTab("marketplace", false, false)).toBe(null);
 });
