+++ frontend/src/components/persona/PersonaEditorModal.tsx
@@ -1,12 +1,11 @@
 import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { GlassSelect } from "../common/GlassSelect";
-import { Plus, Pencil, Sparkles, Tag, Save, MessageSquare, Puzzle, Server } from "lucide-react";
+import { Plus, Pencil, Sparkles, Tag, Save, MessageSquare, Server } from "lucide-react";
 import { LoadingSpinner } from "../common/LoadingSpinner";
 import { EditorSidebar } from "../common/EditorSidebar";
 import toast from "react-hot-toast";
 import { mcpApi } from "../../services/api/mcp";
-import { pluginApi } from "../../services/api/plugin";
 import {
   buildPersonaPresetPayload,
   draftRowsToStarterPrompts,
@@ -51,14 +50,11 @@ export function PersonaEditorModal({
     starter_prompts: starterPromptsToDraftRows(editingPreset?.starter_prompts),
     tags: editingPreset?.tags.join(", ") || "",
     skill_names: [...(editingPreset?.skill_names || [])] as string[],
-    plugin_names: [...(editingPreset?.plugin_names || [])] as string[],
     mcp_server_names: [...(editingPreset?.mcp_server_names || [])] as string[],
   });
 
   const [skillDropdownOpen, setSkillDropdownOpen] = useState(false);
-  const [pluginDropdownOpen, setPluginDropdownOpen] = useState(false);
   const [mcpDropdownOpen, setMcpDropdownOpen] = useState(false);
-  const [pluginOptions, setPluginOptions] = useState<BindingOption[]>([]);
   const [mcpOptions, setMcpOptions] = useState<BindingOption[]>([]);
   const [bindingsLoading, setBindingsLoading] = useState(false);
 
@@ -68,30 +64,18 @@ export function PersonaEditorModal({
     setBindingsLoading(true);
     void (async () => {
       try {
-        const [installs, mcpList] = await Promise.all([
-          pluginApi.listInstalled(),
-          mcpApi.list(),
-        ]);
+        const mcpList = await mcpApi.list();
         if (cancelled) return;
-        setPluginOptions(
-          (installs.installs ?? []).map((record) => ({
-            name: record.plugin_name,
-            description: null,
-          })),
-        );
         setMcpOptions(
           (mcpList.servers ?? [])
             .filter((server) => server.enabled)
             .map((server) => ({
               name: server.name,
-              description: server.source_plugin
-                ? t("personaPresets.mcpFromPlugin", { name: server.source_plugin })
-                : null,
+              description: null,
             })),
         );
       } catch {
         if (!cancelled) {
-          setPluginOptions([]);
           setMcpOptions([]);
         }
       } finally {
@@ -120,8 +104,7 @@ export function PersonaEditorModal({
         ),
         tags: editingPreset?.tags.join(", ") || "",
         skill_names: [...(editingPreset?.skill_names || [])] as string[],
-        plugin_names: [...(editingPreset?.plugin_names || [])] as string[],
-        mcp_server_names: [...(editingPreset?.mcp_server_names || [])] as string[],
+            mcp_server_names: [...(editingPreset?.mcp_server_names || [])] as string[],
       });
       setSkillDropdownOpen(false);
     }
@@ -140,7 +123,6 @@ export function PersonaEditorModal({
         .map((s) => s.trim())
         .filter(Boolean),
       skill_names: draft.skill_names,
-      plugin_names: draft.plugin_names,
       mcp_server_names: draft.mcp_server_names,
     };
 
@@ -416,31 +398,6 @@ export function PersonaEditorModal({
               }
               open={skillDropdownOpen}
               onOpenChange={setSkillDropdownOpen}
-            />
-          </div>
-
-          <div className="ppe-field">
-            <label className="ppe-label">
-              <Puzzle size={13} className="ppe-label-icon" />
-              {t("personaPresets.pluginsInput", "插件")}
-            </label>
-            <PersonaEditorBindingSelector
-              options={pluginOptions}
-              selected={draft.plugin_names}
-              onChange={(updater) =>
-                setDraft((prev) => ({
-                  ...prev,
-                  plugin_names: updater(prev.plugin_names),
-                }))
-              }
-              open={pluginDropdownOpen}
-              onOpenChange={setPluginDropdownOpen}
-              icon={<Puzzle size={12} />}
-              countLabelKey="personaPresets.pluginCount"
-              placeholderKey="personaPresets.pluginsInputPlaceholder"
-              searchPlaceholderKey="personaPresets.pluginsSearchPlaceholder"
-              emptyKey="personaPresets.noPlugins"
-              loading={bindingsLoading}
             />
           </div>
