+++ frontend/src/components/persona/PersonaEditorTypes.ts
@@ -20,7 +20,6 @@ export interface PersonaEditorDraft {
   starter_prompts: { icon: string; text: string }[];
   tags: string;
   skill_names: string[];
-  plugin_names: string[];
   mcp_server_names: string[];
 }
