+++ frontend/src/components/selectors/SkillSelector.tsx
@@ -3,7 +3,6 @@ import { useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import toast from "react-hot-toast";
 import {
-  Puzzle,
   Sparkles,
   ChevronRight,
   Plus,
@@ -51,13 +50,11 @@ interface SkillSelectorProps {
 const sourceIcons: Record<SkillSource, typeof FileCode> = {
   marketplace: Store,
   manual: FileCode,
-  plugin: Puzzle,
 };
 
 const sourceColors: Record<SkillSource, string> = {
   marketplace: "text-[var(--theme-primary)]",
   manual: "text-[var(--theme-text)]",
-  plugin: "text-[var(--theme-primary)]",
 };
 
 export function SkillSelector({
@@ -81,7 +78,7 @@ export function SkillSelector({
   const setIsOpen = externalOnOpenChange ?? setInternalOpen;
   const [expandedCategories, setExpandedCategories] = useState<
     Set<SkillSource>
-  >(new Set(["marketplace", "manual", "plugin"]));
+  >(new Set(["marketplace", "manual"]));
   const [searchQuery, setSearchQuery] = useState("");
   const [selectedTags, setSelectedTags] = useState<string[]>([]);
   const swipeRef = useSwipeToClose({
