+++ frontend/src/i18n/locales/en.json
@@ -679,15 +679,6 @@
     "daemonVersionUnsupported": "Daemon version {{version}} is below minimum {{min}}; please update",
     "sandboxMachineOffline": "Selected sandbox machine '{{machine}}' is offline",
     "sandboxMachineNotFound": "Sandbox machine '{{machine}}' not found or removed",
-    "pluginNotFound": "Plugin {{name}} not found",
-    "pluginInactive": "Plugin {{name}} is not active",
-    "pluginAlreadyInstalled": "Plugin {{name}} already installed",
-    "pluginNotInstalled": "Plugin {{name}} not installed",
-    "pluginNameExists": "Plugin name {{name}} already taken",
-    "pluginMcpNameConflict": "MCP server name {{server}} is already used by another plugin or server",
-    "pluginMcpRefNotFound": "Referenced system MCP server {{server}} not found",
-    "pluginInvalidPayload": "Plugin payload invalid: {{reason}}",
-    "pluginFileNotFound": "File {{path}} not found in plugin {{name}}",
     "pluginNoEditPermission": "No permission to edit plugin {{name}}"
   },
   "categories": {
@@ -1791,63 +1782,6 @@
     "viewInMySkills": "Open in My Skills"
   },
 
-  "plugins": {
-"title": "Plugin Market",
-"subtitle": "Install platform plugins to get skills, MCP servers and personas at once",
-"searchPlaceholder": "Search plugins...",
-"noPlugins": "No plugins yet",
-"noMatchingPlugins": "No matching plugins",
-"emptyHint": "Plugins are curated by platform admins; installing brings their skills and tools automatically",
-"noDescription": "No description",
-"installed": "Installed",
-"draft": "Draft",
-"inactive": "Deactivated",
-"install": "Install",
-"update": "Update",
-"uninstall": "Uninstall",
-"preview": "Preview",
-"activate": "Activate",
-"deactivate": "Deactivate",
-"installCount_one": "{{count}} install",
-"installCount_other": "{{count}} installs",
-"capabilitySkills_one": "{{count}} skill",
-"capabilitySkills_other": "{{count}} skills",
-"capabilityMcp_one": "{{count}} MCP",
-"capabilityMcp_other": "{{count}} MCP",
-"capabilityPersona": "Persona",
-"confirmInstall": "Install plugin \"{{name}}\"?",
-"confirmInstallMessage": "Skills bundled in this plugin will be installed into your library; the persona preset is copied as a private draft.",
-"confirmUpdate": "Update plugin \"{{name}}\"?",
-"confirmUpdateMessage": "The latest version will be pulled from the platform and overwrite local skill copies.",
-"confirmUninstall": "Uninstall plugin \"{{name}}\"?",
-"confirmUninstallMessage": "Only the install record is removed; installed skills and persona copies stay unchanged.",
-"confirmDelete": "Delete plugin \"{{name}}\"?",
-"confirmDeleteMessage": "The plugin and its materialized MCP servers will be removed; installed copies are unaffected.",
-"installSuccess": "Plugin {{name}} installed",
-"installFailed": "Failed to install plugin",
-"updateSuccess": "Plugin {{name}} updated",
-"updateFailed": "Failed to update plugin",
-"uninstallSuccess": "Plugin {{name}} uninstalled",
-"uninstallFailed": "Failed to uninstall plugin",
-"activateSuccess": "Plugin activated",
-"deactivateSuccess": "Plugin deactivated",
-"deleteSuccess": "Plugin deleted",
-"deleteFailed": "Failed to delete plugin",
-"fetchFailed": "Failed to load plugin market",
-"fetchTagsFailed": "Failed to load tags:",
-"clearFilters": "Clear filters",
-"filterAll": "All",
-"filterActive": "Active",
-"filterInactive": "Inactive",
-"payloadSkills": "Skills",
-"payloadMcp": "MCP Servers",
-"payloadPersona": "Persona",
-"noSkills": "This plugin bundles no skills",
-"noMcp": "This plugin bundles no MCP servers",
-"mcpRefBadge": "References platform server",
-"filePreview": "File preview",
-"previewLoadFailed": "Failed to read file"
-  },
   "mcp": {
     "addFirst": "Add your first MCP server",
     "addNew": "Add New MCP Server",
@@ -1858,7 +1792,6 @@
       "disable": "Disable",
       "disableTool": "Disable tool",
       "disabled": "Disabled",
-      "fromPlugin": "From plugin {{name}}",
       "discoverFailed": "Failed to discover tools",
       "discovering": "Discovering tools...",
       "edit": "Edit",
@@ -2048,7 +1981,6 @@
     "groupFeature": "Features",
     "groupSystem": "System",
     "marketplace": "Marketplace",
-    "plugins": "Plugins",
     "mcp": "MCP Servers",
     "memory": "Memory Space",
     "models": "Models",
@@ -2209,11 +2141,6 @@
     "skillsControlledHint": "The current persona \"{{name}}\" controls the available Skills. Clear the persona or edit the preset to change skills for this chat.",
     "skillsCount": "skills",
     "skillsInput": "Skills",
-    "pluginsInput": "Plugins",
-    "pluginCount": "{{count}} plugins bound",
-    "pluginsInputPlaceholder": "Bind installed plugins...",
-    "pluginsSearchPlaceholder": "Search plugins...",
-    "noPlugins": "No installed plugins",
     "mcpServersInput": "MCP Servers",
     "mcpServerCount": "{{count}} MCP servers bound",
     "mcpServersInputPlaceholder": "Restrict usable MCP servers (empty = all)...",
@@ -2632,10 +2559,6 @@
     "marketplace": {
       "description": "Discover and install community-built AI skills. Browse, rate, and deploy pre-built capabilities to extend your AI agent functionality.",
       "title": "Skill Marketplace"
-    },
-    "plugins": {
-      "description": "Install platform plugins to get skills, MCP servers and personas at once.",
-      "title": "Plugin Hub"
     },
     "mcp": {
       "description": "Manage Model Context Protocol server connections. Connect external tools, databases, and services to extend AI agent capabilities.",
@@ -3230,7 +3153,6 @@
       "${cat}": "{{cat}}",
       "github": "GitHub",
       "manual": "Manual",
-      "plugin": "Plugin",
       "marketplace": "Marketplace"
     },
     "system": "System",
