+++ frontend/src/i18n/locales/ja.json
@@ -679,15 +679,6 @@
     "daemonVersionUnsupported": "ローカルサンドボックスデーモンのバージョン {{version}} は最低要件 {{min}} 未満です。アップデートしてください",
     "sandboxMachineOffline": "選択したサンドボックスマシン「{{machine}}」は現在オフラインです",
     "sandboxMachineNotFound": "サンドボックスマシン「{{machine}}」が見つからないか、削除されました",
-    "pluginNotFound": "プラグイン {{name}} が見つかりません",
-    "pluginInactive": "プラグイン {{name}} は有効ではありません",
-    "pluginAlreadyInstalled": "プラグイン {{name}} は既にインストール済みです",
-    "pluginNotInstalled": "プラグイン {{name}} はインストールされていません",
-    "pluginNameExists": "プラグイン名 {{name}} は既に使用されています",
-    "pluginMcpNameConflict": "MCP サーバー名 {{server}} は他のプラグインまたはサーバーで使用中です",
-    "pluginMcpRefNotFound": "参照先のシステム MCP サーバー {{server}} が見つかりません",
-    "pluginInvalidPayload": "プラグインのペイロードが無効です: {{reason}}",
-    "pluginFileNotFound": "プラグイン {{name}} にファイル {{path}} が見つかりません",
     "pluginNoEditPermission": "プラグイン {{name}} を編集する権限がありません"
   },
   "categories": {
@@ -1791,63 +1782,6 @@
     "viewInMySkills": "マイスキルで開く"
   },
 
-  "plugins": {
-"title": "プラグインマーケット",
-"subtitle": "プラットフォームプラグインをインストールして、スキル・MCP サーバー・ペルソナを一度に入手",
-"searchPlaceholder": "プラグインを検索...",
-"noPlugins": "プラグインはまだありません",
-"noMatchingPlugins": "一致するプラグインがありません",
-"emptyHint": "プラグインはプラットフォーム管理者が公開します。インストールするとスキルとツールが自動的に追加されます",
-"noDescription": "説明なし",
-"installed": "インストール済み",
-"draft": "ドラフト",
-"inactive": "無効",
-"install": "インストール",
-"update": "更新",
-"uninstall": "アンインストール",
-"preview": "プレビュー",
-"activate": "有効化",
-"deactivate": "無効化",
-"installCount_one": "{{count}} 回インストール",
-"installCount_other": "{{count}} 回インストール",
-"capabilitySkills_one": "スキル {{count}} 件",
-"capabilitySkills_other": "スキル {{count}} 件",
-"capabilityMcp_one": "MCP {{count}} 件",
-"capabilityMcp_other": "MCP {{count}} 件",
-"capabilityPersona": "ペルソナ付き",
-"confirmInstall": "プラグイン「{{name}}」をインストールしますか？",
-"confirmInstallMessage": "プラグインのスキルがライブラリに追加され、ペルソナは下書きとしてコピーされます。",
-"confirmUpdate": "プラグイン「{{name}}」を更新しますか？",
-"confirmUpdateMessage": "最新版を取得し、ローカルのスキルを上書きします。",
-"confirmUninstall": "プラグイン「{{name}}」をアンインストールしますか？",
-"confirmUninstallMessage": "インストール記録のみ削除されます。スキルとペルソナのコピーは残ります。",
-"confirmDelete": "プラグイン「{{name}}」を削除しますか？",
-"confirmDeleteMessage": "プラグインと生成された MCP サーバーが削除されます。インストール済みコピーには影響しません。",
-"installSuccess": "プラグイン {{name}} をインストールしました",
-"installFailed": "プラグインのインストールに失敗しました",
-"updateSuccess": "プラグイン {{name}} を更新しました",
-"updateFailed": "プラグインの更新に失敗しました",
-"uninstallSuccess": "プラグイン {{name}} をアンインストールしました",
-"uninstallFailed": "プラグインのアンインストールに失敗しました",
-"activateSuccess": "プラグインを有効化しました",
-"deactivateSuccess": "プラグインを無効化しました",
-"deleteSuccess": "プラグインを削除しました",
-"deleteFailed": "プラグインの削除に失敗しました",
-"fetchFailed": "プラグインマーケットの取得に失敗しました",
-"fetchTagsFailed": "タグの取得に失敗:",
-"clearFilters": "フィルターを解除",
-"filterAll": "すべて",
-"filterActive": "有効",
-"filterInactive": "無効",
-"payloadSkills": "スキル",
-"payloadMcp": "MCP サーバー",
-"payloadPersona": "ペルソナ",
-"noSkills": "このプラグインにスキルは含まれません",
-"noMcp": "このプラグインに MCP サーバーは含まれません",
-"mcpRefBadge": "プラットフォームサーバーを参照",
-"filePreview": "ファイルプレビュー",
-"previewLoadFailed": "ファイルの読み込みに失敗しました"
-  },
   "mcp": {
     "addFirst": "最初のMCPサーバーを追加",
     "addNew": "新しいMCPサーバーを追加",
@@ -1858,7 +1792,6 @@
       "disable": "無効化",
       "disableTool": "ツールを無効化",
       "disabled": "無効",
-      "fromPlugin": "プラグイン {{name}} から",
       "discoverFailed": "ツールの検出に失敗しました",
       "discovering": "ツールを検出中...",
       "edit": "編集",
@@ -2048,7 +1981,6 @@
     "groupFeature": "機能",
     "groupSystem": "システム",
     "marketplace": "マーケットプレイス",
-    "plugins": "プラグイン",
     "mcp": "MCP サービス",
     "memory": "記憶スペース",
     "models": "モデル",
@@ -2209,11 +2141,6 @@
     "skillsControlledHint": "現在のペルソナ「{{name}}」が使用可能なスキルを制御しています。スキルを変更するには、ペルソナを解除するかプリセットを編集してください。",
     "skillsCount": "スキル",
     "skillsInput": "スキル",
-    "pluginsInput": "プラグイン",
-    "pluginCount": "プラグイン {{count}} 件をバインド",
-    "pluginsInputPlaceholder": "インストール済みプラグインをバインド...",
-    "pluginsSearchPlaceholder": "プラグインを検索...",
-    "noPlugins": "インストール済みプラグインはありません",
     "mcpServersInput": "MCP サーバー",
     "mcpServerCount": "MCP サーバー {{count}} 件をバインド",
     "mcpServersInputPlaceholder": "使用する MCP サーバーを限定（空欄で無制限）...",
@@ -2632,10 +2559,6 @@
     "marketplace": {
       "description": "コミュニティ構築のAIスキルを発見・インストール。評価・デプロイでAIエージェントの機能を拡張。",
       "title": "スキルマーケットプレイス"
-    },
-    "plugins": {
-      "description": "プラットフォームプラグインをインストールして、スキル・MCP・ペルソナを一度に入手。",
-      "title": "プラグインセンター"
     },
     "mcp": {
       "description": "Model Context Protocolサーバー接続の管理。外部ツール、データベース、サービスを接続してAIエージェントの能力を拡張。",
@@ -3230,7 +3153,6 @@
       "${cat}": "{{cat}}",
       "github": "GitHub",
       "manual": "手動作成",
-      "plugin": "プラグイン",
       "marketplace": "マーケット"
     },
     "system": "システム",
