+++ frontend/src/i18n/locales/ko.json
@@ -679,15 +679,6 @@
     "daemonVersionUnsupported": "로컬 샌드박스 데몬 버전 {{version}}이(가) 최소 요구 사항 {{min}} 미만입니다. 업데이트해 주세요",
     "sandboxMachineOffline": "선택한 샌드박스 머신 '{{machine}}'이(가) 오프라인 상태입니다",
     "sandboxMachineNotFound": "샌드박스 머신 '{{machine}}'을(를) 찾을 수 없거나 제거되었습니다",
-    "pluginNotFound": "플러그인 {{name}}을(를) 찾을 수 없습니다",
-    "pluginInactive": "플러그인 {{name}}이(가) 활성화되지 않았습니다",
-    "pluginAlreadyInstalled": "플러그인 {{name}}이(가) 이미 설치되어 있습니다",
-    "pluginNotInstalled": "플러그인 {{name}}이(가) 설치되어 있지 않습니다",
-    "pluginNameExists": "플러그인 이름 {{name}}이(가) 이미 사용 중입니다",
-    "pluginMcpNameConflict": "MCP 서버 이름 {{server}}은(는) 이미 다른 플러그인이나 서버에서 사용 중입니다",
-    "pluginMcpRefNotFound": "참조한 시스템 MCP 서버 {{server}}를 찾을 수 없습니다",
-    "pluginInvalidPayload": "플러그인 페이로드가 잘못되었습니다: {{reason}}",
-    "pluginFileNotFound": "플러그인 {{name}}에서 파일 {{path}}을(를) 찾을 수 없습니다",
     "pluginNoEditPermission": "플러그인 {{name}}을(를) 편집할 권한이 없습니다"
   },
   "categories": {
@@ -1791,63 +1782,6 @@
     "viewInMySkills": "내 스킬에서 열기"
   },
 
-  "plugins": {
-"title": "플러그인 마켓",
-"subtitle": "플랫폼 플러그인을 설치해 스킬·MCP 서버·페르소나를 한 번에 확보",
-"searchPlaceholder": "플러그인 검색...",
-"noPlugins": "플러그인이 없습니다",
-"noMatchingPlugins": "일치하는 플러그인이 없습니다",
-"emptyHint": "플러그인은 플랫폼 관리자가 등록합니다. 설치하면 스킬과 도구가 자동으로 추가됩니다",
-"noDescription": "설명 없음",
-"installed": "설치됨",
-"draft": "초안",
-"inactive": "비활성화",
-"install": "설치",
-"update": "업데이트",
-"uninstall": "제거",
-"preview": "미리보기",
-"activate": "활성화",
-"deactivate": "비활성화",
-"installCount_one": "설치 {{count}}회",
-"installCount_other": "설치 {{count}}회",
-"capabilitySkills_one": "스킬 {{count}}개",
-"capabilitySkills_other": "스킬 {{count}}개",
-"capabilityMcp_one": "MCP {{count}}개",
-"capabilityMcp_other": "MCP {{count}}개",
-"capabilityPersona": "페르소나 포함",
-"confirmInstall": "플러그인 \"{{name}}\"을(를) 설치할까요?",
-"confirmInstallMessage": "플러그인의 스킬이 라이브러리에 추가되고, 페르소나는 초안으로 복사됩니다.",
-"confirmUpdate": "플러그인 \"{{name}}\"을(를) 업데이트할까요?",
-"confirmUpdateMessage": "최신 버전을 가져와 로컬 스킬을 덮어씁니다.",
-"confirmUninstall": "플러그인 \"{{name}}\"을(를) 제거할까요?",
-"confirmUninstallMessage": "설치 기록만 제거되며, 설치된 스킬과 페르소나 사본은 유지됩니다.",
-"confirmDelete": "플러그인 \"{{name}}\"을(를) 삭제할까요?",
-"confirmDeleteMessage": "플러그인과 생성된 MCP 서버가 삭제됩니다. 설치된 사본은 영향을 받지 않습니다.",
-"installSuccess": "플러그인 {{name}} 설치 완료",
-"installFailed": "플러그인 설치 실패",
-"updateSuccess": "플러그인 {{name}} 업데이트 완료",
-"updateFailed": "플러그인 업데이트 실패",
-"uninstallSuccess": "플러그인 {{name}} 제거 완료",
-"uninstallFailed": "플러그인 제거 실패",
-"activateSuccess": "플러그인이 활성화되었습니다",
-"deactivateSuccess": "플러그인이 비활성화되었습니다",
-"deleteSuccess": "플러그인이 삭제되었습니다",
-"deleteFailed": "플러그인 삭제 실패",
-"fetchFailed": "플러그인 마켓을 불러오지 못했습니다",
-"fetchTagsFailed": "태그를 불러오지 못했습니다:",
-"clearFilters": "필터 지우기",
-"filterAll": "전체",
-"filterActive": "활성",
-"filterInactive": "비활성",
-"payloadSkills": "스킬",
-"payloadMcp": "MCP 서버",
-"payloadPersona": "페르소나",
-"noSkills": "이 플러그인에는 스킬이 없습니다",
-"noMcp": "이 플러그인에는 MCP 서버가 없습니다",
-"mcpRefBadge": "플랫폼 서버 참조",
-"filePreview": "파일 미리보기",
-"previewLoadFailed": "파일을 읽지 못했습니다"
-  },
   "mcp": {
     "addFirst": "첫 번째 MCP 서버 추가",
     "addNew": "새 MCP 서버 추가",
@@ -1858,7 +1792,6 @@
       "disable": "비활성화",
       "disableTool": "도구 비활성화",
       "disabled": "비활성화됨",
-      "fromPlugin": "플러그인 {{name}}",
       "discoverFailed": "도구 검색 실패",
       "discovering": "도구 검색 중...",
       "edit": "편집",
@@ -2048,7 +1981,6 @@
     "groupFeature": "기능",
     "groupSystem": "시스템",
     "marketplace": "마켓플레이스",
-    "plugins": "플러그인",
     "mcp": "MCP 서비스",
     "memory": "메모리 공간",
     "models": "모델",
@@ -2209,11 +2141,6 @@
     "skillsControlledHint": "현재 페르소나 \"{{name}}\"이(가) 사용 가능한 스킬을 제어합니다. 이 채팅의 스킬을 변경하려면 페르소나를 해제하거나 프리셋을 편집하세요.",
     "skillsCount": "스킬",
     "skillsInput": "스킬",
-    "pluginsInput": "플러그인",
-    "pluginCount": "플러그인 {{count}}개 바인딩됨",
-    "pluginsInputPlaceholder": "설치된 플러그인 바인딩...",
-    "pluginsSearchPlaceholder": "플러그인 검색...",
-    "noPlugins": "설치된 플러그인이 없습니다",
     "mcpServersInput": "MCP 서버",
     "mcpServerCount": "MCP 서버 {{count}}개 바인딩됨",
     "mcpServersInputPlaceholder": "사용할 MCP 서버 제한(비워두면 전부)...",
@@ -2632,10 +2559,6 @@
     "marketplace": {
       "description": "커뮤니티에서 제작한 AI 스킬을 발견하고 설치하세요. 평가, 배포로 AI 에이전트 기능을 확장합니다.",
       "title": "스킬 마켓플레이스"
-    },
-    "plugins": {
-      "description": "플랫폼 플러그인을 설치해 스킬·MCP·페르소나를 한 번에 확보.",
-      "title": "플러그인 센터"
     },
     "mcp": {
       "description": "Model Context Protocol 서버 연결을 관리하세요. 외부 도구, 데이터베이스 및 서비스를 연결하여 AI 에이전트 기능을 확장합니다.",
@@ -3230,7 +3153,6 @@
       "${cat}": "{{cat}}",
       "github": "GitHub",
       "manual": "수동 생성",
-      "plugin": "플러그인",
       "marketplace": "마켓플레이스"
     },
     "system": "시스템",
