+++ frontend/src/i18n/locales/ru.json
@@ -679,15 +679,6 @@
     "daemonVersionUnsupported": "Версия локального sandbox-демона {{version}} ниже минимальной {{min}}, обновите его",
     "sandboxMachineOffline": "Выбранная машина песочницы «{{machine}}» сейчас не в сети",
     "sandboxMachineNotFound": "Машина песочницы «{{machine}}» не найдена или удалена",
-    "pluginNotFound": "Плагин {{name}} не найден",
-    "pluginInactive": "Плагин {{name}} не активен",
-    "pluginAlreadyInstalled": "Плагин {{name}} уже установлен",
-    "pluginNotInstalled": "Плагин {{name}} не установлен",
-    "pluginNameExists": "Имя плагина {{name}} уже занято",
-    "pluginMcpNameConflict": "Имя MCP-сервера {{server}} уже используется другим плагином или сервером",
-    "pluginMcpRefNotFound": "Ссылка на системный MCP-сервер {{server}} не найдена",
-    "pluginInvalidPayload": "Некорректная полезная нагрузка плагина: {{reason}}",
-    "pluginFileNotFound": "Файл {{path}} не найден в плагине {{name}}",
     "pluginNoEditPermission": "Нет прав на редактирование плагина {{name}}"
   },
   "categories": {
@@ -1791,63 +1782,6 @@
     "viewInMySkills": "Открыть в моих навыках"
   },
 
-  "plugins": {
-"title": "Маркет плагинов",
-"subtitle": "Устанавливайте плагины платформы: навыки, MCP-серверы и персоны сразу",
-"searchPlaceholder": "Поиск плагинов...",
-"noPlugins": "Плагинов пока нет",
-"noMatchingPlugins": "Нет подходящих плагинов",
-"emptyHint": "Плагины публикуются администраторами платформы; установка добавляет их навыки и инструменты",
-"noDescription": "Без описания",
-"installed": "Установлен",
-"draft": "Черновик",
-"inactive": "Отключён",
-"install": "Установить",
-"update": "Обновить",
-"uninstall": "Удалить",
-"preview": "Просмотр",
-"activate": "Включить",
-"deactivate": "Отключить",
-"installCount_one": "{{count}} установка",
-"installCount_other": "{{count}} установок",
-"capabilitySkills_one": "{{count}} навык",
-"capabilitySkills_other": "{{count}} навыков",
-"capabilityMcp_one": "{{count}} MCP",
-"capabilityMcp_other": "{{count}} MCP",
-"capabilityPersona": "С персоной",
-"confirmInstall": "Установить плагин «{{name}}»?",
-"confirmInstallMessage": "Навыки плагина будут добавлены в вашу библиотеку, персона скопируется как черновик.",
-"confirmUpdate": "Обновить плагин «{{name}}»?",
-"confirmUpdateMessage": "Последняя версия будет загружена и перезапишет локальные копии навыков.",
-"confirmUninstall": "Удалить плагин «{{name}}»?",
-"confirmUninstallMessage": "Удаляется только запись об установке; навыки и копии персон сохраняются.",
-"confirmDelete": "Удалить плагин «{{name}}»?",
-"confirmDeleteMessage": "Плагин и его MCP-серверы будут удалены; установленные копии не затрагиваются.",
-"installSuccess": "Плагин {{name}} установлен",
-"installFailed": "Не удалось установить плагин",
-"updateSuccess": "Плагин {{name}} обновлён",
-"updateFailed": "Не удалось обновить плагин",
-"uninstallSuccess": "Плагин {{name}} удалён",
-"uninstallFailed": "Не удалось удалить плагин",
-"activateSuccess": "Плагин включён",
-"deactivateSuccess": "Плагин отключён",
-"deleteSuccess": "Плагин удалён",
-"deleteFailed": "Не удалось удалить плагин",
-"fetchFailed": "Не удалось загрузить маркет плагинов",
-"fetchTagsFailed": "Не удалось загрузить теги:",
-"clearFilters": "Сбросить фильтры",
-"filterAll": "Все",
-"filterActive": "Активные",
-"filterInactive": "Отключённые",
-"payloadSkills": "Навыки",
-"payloadMcp": "MCP-серверы",
-"payloadPersona": "Персона",
-"noSkills": "Плагин не содержит навыков",
-"noMcp": "Плагин не содержит MCP-серверов",
-"mcpRefBadge": "Ссылка на сервер платформы",
-"filePreview": "Просмотр файла",
-"previewLoadFailed": "Не удалось прочитать файл"
-  },
   "mcp": {
     "addFirst": "Добавьте первый MCP сервер",
     "addNew": "Добавить новый MCP сервер",
@@ -1858,7 +1792,6 @@
       "disable": "Отключить",
       "disableTool": "Отключить инструмент",
       "disabled": "Отключён",
-      "fromPlugin": "Из плагина {{name}}",
       "discoverFailed": "Не удалось обнаружить инструменты",
       "discovering": "Обнаружение инструментов...",
       "edit": "Редактировать",
@@ -2048,7 +1981,6 @@
     "groupFeature": "Функции",
     "groupSystem": "Система",
     "marketplace": "Магазин",
-    "plugins": "Плагины",
     "mcp": "MCP Серверы",
     "memory": "Пространство памяти",
     "models": "Модели",
@@ -2209,11 +2141,6 @@
     "skillsControlledHint": "Текущая персона \"{{name}}\" управляет доступными навыками. Очистите персону или измените пресет, чтобы сменить навыки для этого чата.",
     "skillsCount": "навыков",
     "skillsInput": "Навыки",
-    "pluginsInput": "Плагины",
-    "pluginCount": "Привязано плагинов: {{count}}",
-    "pluginsInputPlaceholder": "Привязать установленные плагины...",
-    "pluginsSearchPlaceholder": "Поиск плагинов...",
-    "noPlugins": "Нет установленных плагинов",
     "mcpServersInput": "MCP-серверы",
     "mcpServerCount": "Привязано MCP-серверов: {{count}}",
     "mcpServersInputPlaceholder": "Ограничить MCP-серверы (пусто = все)...",
@@ -2632,10 +2559,6 @@
     "marketplace": {
       "description": "Открывайте и устанавливайте AI-навыки от сообщества. Просматривайте, оценивайте и развёртывайте готовые функции для расширения возможностей AI-агента.",
       "title": "Маркетплейс навыков"
-    },
-    "plugins": {
-      "description": "Устанавливайте плагины платформы: навыки, MCP-серверы и персоны сразу.",
-      "title": "Центр плагинов"
     },
     "mcp": {
       "description": "Управляйте подключениями серверов Model Context Protocol. Подключайте внешние инструменты, базы данных и сервисы для расширения возможностей AI-агента.",
@@ -3230,7 +3153,6 @@
       "${cat}": "{{cat}}",
       "github": "GitHub",
       "manual": "Ручной",
-      "plugin": "Плагин",
       "marketplace": "Магазин"
     },
     "system": "Системный",
