+++ frontend/src/i18n/locales/zh.json
@@ -679,15 +679,6 @@
     "daemonVersionUnsupported": "本地沙箱守护进程版本 {{version}} 低于最低要求 {{min}}，请升级",
     "sandboxMachineOffline": "所选沙箱机器「{{machine}}」当前离线",
     "sandboxMachineNotFound": "沙箱机器「{{machine}}」不存在或已移除",
-    "pluginNotFound": "插件 {{name}} 不存在",
-    "pluginInactive": "插件 {{name}} 未激活",
-    "pluginAlreadyInstalled": "插件 {{name}} 已安装",
-    "pluginNotInstalled": "插件 {{name}} 尚未安装",
-    "pluginNameExists": "插件名 {{name}} 已被占用",
-    "pluginMcpNameConflict": "MCP 服务名 {{server}} 已被其他插件或服务占用",
-    "pluginMcpRefNotFound": "引用的平台 MCP 服务 {{server}} 不存在",
-    "pluginInvalidPayload": "插件负载无效：{{reason}}",
-    "pluginFileNotFound": "插件 {{name}} 中找不到文件 {{path}}",
     "pluginNoEditPermission": "没有编辑插件 {{name}} 的权限"
   },
   "categories": {
@@ -1791,63 +1782,6 @@
     "viewInMySkills": "去我的技能查看"
   },
 
-  "plugins": {
-"title": "插件市场",
-"subtitle": "安装平台插件，一次获得技能、MCP 服务与角色预设",
-"searchPlaceholder": "搜索插件...",
-"noPlugins": "暂无插件",
-"noMatchingPlugins": "没有匹配的插件",
-"emptyHint": "插件由平台管理员上架，安装后自动获得其中的技能与工具",
-"noDescription": "暂无描述",
-"installed": "已安装",
-"draft": "草稿",
-"inactive": "已停用",
-"install": "安装",
-"update": "更新",
-"uninstall": "卸载",
-"preview": "预览",
-"activate": "激活",
-"deactivate": "停用",
-"installCount_one": "{{count}} 次安装",
-"installCount_other": "{{count}} 次安装",
-"capabilitySkills_one": "{{count}} 个技能",
-"capabilitySkills_other": "{{count}} 个技能",
-"capabilityMcp_one": "{{count}} 个 MCP",
-"capabilityMcp_other": "{{count}} 个 MCP",
-"capabilityPersona": "含角色",
-"confirmInstall": "安装插件「{{name}}」？",
-"confirmInstallMessage": "将把插件携带的技能安装到你的技能库，角色预设复制为私有草稿。",
-"confirmUpdate": "更新插件「{{name}}」？",
-"confirmUpdateMessage": "将从平台拉取最新版本并覆盖本地技能副本。",
-"confirmUninstall": "卸载插件「{{name}}」？",
-"confirmUninstallMessage": "仅移除安装记录；已安装的技能与角色副本保持不变。",
-"confirmDelete": "删除插件「{{name}}」？",
-"confirmDeleteMessage": "将删除插件及其物化的 MCP 服务，已安装副本不受影响。",
-"installSuccess": "插件 {{name}} 安装成功",
-"installFailed": "安装插件失败",
-"updateSuccess": "插件 {{name}} 已更新",
-"updateFailed": "更新插件失败",
-"uninstallSuccess": "插件 {{name}} 已卸载",
-"uninstallFailed": "卸载插件失败",
-"activateSuccess": "插件已激活",
-"deactivateSuccess": "插件已停用",
-"deleteSuccess": "插件已删除",
-"deleteFailed": "删除插件失败",
-"fetchFailed": "获取插件市场失败",
-"fetchTagsFailed": "获取标签失败:",
-"clearFilters": "清除筛选",
-"filterAll": "全部",
-"filterActive": "已激活",
-"filterInactive": "已停用",
-"payloadSkills": "技能负载",
-"payloadMcp": "MCP 服务",
-"payloadPersona": "角色负载",
-"noSkills": "该插件不携带技能",
-"noMcp": "该插件不携带 MCP 服务",
-"mcpRefBadge": "引用平台服务",
-"filePreview": "文件预览",
-"previewLoadFailed": "读取文件失败"
-  },
   "mcp": {
     "addFirst": "添加您的第一个 MCP 服务器",
     "addNew": "添加新 MCP 服务器",
@@ -1858,7 +1792,6 @@
       "disable": "禁用",
       "disableTool": "禁用工具",
       "disabled": "已禁用",
-      "fromPlugin": "来自插件 {{name}}",
       "discoverFailed": "发现工具失败",
       "discovering": "正在发现工具...",
       "edit": "编辑",
@@ -2048,7 +1981,6 @@
     "groupFeature": "功能",
     "groupSystem": "系统",
     "marketplace": "技能商店",
-    "plugins": "插件中心",
     "mcp": "MCP 服务",
     "memory": "记忆空间",
     "models": "模型",
@@ -2209,11 +2141,6 @@
     "skillsControlledHint": "当前角色「{{name}}」正在控制可用 Skills。要调整本次对话的技能，请先清除当前角色，或编辑该角色预设。",
     "skillsCount": "个技能",
     "skillsInput": "技能",
-    "pluginsInput": "插件",
-    "pluginCount": "{{count}} 个插件已绑定",
-    "pluginsInputPlaceholder": "绑定已安装的插件...",
-    "pluginsSearchPlaceholder": "搜索插件...",
-    "noPlugins": "暂无已安装的插件",
     "mcpServersInput": "MCP 服务",
     "mcpServerCount": "{{count}} 个 MCP 服务已绑定",
     "mcpServersInputPlaceholder": "限定可用的 MCP 服务（留空不限制）...",
@@ -2632,10 +2559,6 @@
     "marketplace": {
       "description": "发现并安装社区构建的 AI 技能。浏览、评分和部署预构建能力，扩展 AI 智能体功能。",
       "title": "技能市场"
-    },
-    "plugins": {
-      "description": "安装平台插件，一次获得技能、MCP 服务与角色预设。",
-      "title": "插件中心"
     },
     "mcp": {
       "description": "管理模型上下文协议服务器连接。连接外部工具、数据库和服务，扩展 AI 智能体能力。",
@@ -3230,7 +3153,6 @@
       "${cat}": "{{cat}}",
       "github": "GitHub",
       "manual": "手动创建",
-      "plugin": "插件",
       "marketplace": "技能商店安装"
     },
     "system": "系统",
