+++ frontend/src/types/mcp.ts
@@ -16,7 +16,6 @@ export interface MCPServerBase {
 
 // MCP Server Response (from API)
 export interface MCPServerResponse extends MCPServerBase {
-  source_plugin?: string | null;
   is_system: boolean;
   is_internal?: boolean;
   can_edit: boolean;
