+++ frontend/src/types/personaPreset.ts
@@ -19,7 +19,6 @@ export interface PersonaPreset {
   system_prompt: string;
   starter_prompts?: PersonaStarterPrompt[];
   skill_names: string[];
-  plugin_names: string[];
   mcp_server_names: string[];
   visibility: PersonaPresetVisibility;
   status: PersonaPresetStatus;
@@ -44,7 +43,6 @@ export interface PersonaPresetCreate {
   system_prompt: string;
   starter_prompts?: PersonaStarterPrompt[];
   skill_names?: string[];
-  plugin_names?: string[];
   mcp_server_names?: string[];
   scope?: PersonaPresetScope;
   visibility?: PersonaPresetVisibility;
@@ -59,7 +57,6 @@ export interface PersonaPresetUpdate {
   system_prompt?: string;
   starter_prompts?: PersonaStarterPrompt[];
   skill_names?: string[];
-  plugin_names?: string[];
   mcp_server_names?: string[];
   scope?: PersonaPresetScope;
   visibility?: PersonaPresetVisibility;
@@ -78,8 +75,6 @@ export interface PersonaPresetSnapshot {
   starter_prompts?: PersonaStarterPrompt[];
   skill_names: string[];
   missing_skill_names: string[];
-  plugin_names: string[];
-  missing_plugin_names: string[];
   mcp_server_names: string[];
   version: number;
   avatar?: string | null;
