+++ src/infra/skill/types.py
@@ -9,7 +9,6 @@ class InstalledFrom(str, Enum):
 
     MARKETPLACE = "marketplace"
     MANUAL = "manual"
-    PLUGIN = "plugin"
 
 
 class MarketplaceSkill(BaseModel):
